Created project and task controllers and fixed some service errors

+++ src/main/java/com/example/demo/services/TasksServices.java
@@ -2,14 +2,16 @@ package com.example.demo.services;
 
 import java.util.List;
 
-import com.example.demo.entity.Project;
 import com.example.demo.entity.Task;
 
 public interface TasksServices {
-	
+
 	List<Task> showAllTasks();
+
 	Task addTask(Task task);
-	int deleteTask(int id);
+
+	int deleteTask(Long id);
+
 	Task updateTask(Task task);
 
 }
